修复 bug

+++ plugins/handlers/message.py
@@ -179,7 +179,8 @@ def verify_ask(client: Client, message: Message) -> bool:
             wait_group_list = list(glovar.user_ids[uid]["wait"])
 
             # Check wait list
-            if not wait_group_list or all(is_should_qns(g) for g in wait_group_list):
+            if not wait_group_list or all(is_should_qns(g) and glovar.user_ids[uid]["qns"].get(g)
+                                          for g in wait_group_list):
                 kick_user(client, gid, uid)
                 delete_message(client, gid, mid)
                 continue
@@ -236,7 +237,8 @@ def verify_check(client: Client, message: Message) -> bool:
         wait_group_list = list(glovar.user_ids[uid]["wait"])
 
         # Check wait list
-        if not wait_group_list or all(is_should_qns(g) for g in wait_group_list):
+        if not wait_group_list or all(is_should_qns(g) and glovar.user_ids[uid]["qns"].get(g)
+                                      for g in wait_group_list):
             return False
 
         # Check if the user is Class D personnel
